Reboot the repository
Starting this new year, clean the repository and start anew. Scrap as
much code as possible and try to reuse and improve upon it. Yet send
everything that is not useful to the recycle bin.

Most files have been sent to the folder `2021-22`. This folder will
store old, cumbersome scripts and files.