Update transcripcio.py
Measure the time it takes to run the script with the datetime module.

Create a "Printing" class that prints STATUS updates in green and
WARNING messages in yellow, prepending the date and hour in which they
were printed.

Small changes to some comments.

Within the Mapping class, identify minor scaffolds and chromosomes with
a pattern. Moreover, add "Q." and "T." to Qnames and Tnames.

Start creating list_chromosomes and chromosome_coverage methods of
Mapping class.