Update transcripcio.py
Remove obsolete comments.

Create a dictionary with sequence sizes. Sadly, it is considerably slow
to create...

Create a list of distances between mappings, alignment lengths for query
and target, and other miscellaneous variables. Uses a hidden function
(starting with '_').

I am working towards defining a function that reads a list of `Mapping`
objects and produces a table from them!