Update pafparser.py
Improve Mapping.__init__ documentation.

Remove redundant subsets of the main dataframe (dataframes where sequids
are only chromosomes). They can be obtained via loc access when
necessary, freeing up memory usage.
In its place, create a column indicating the type of sequid subset of
the row (dtype categorical).

New setting of Mapping.__init__: `memory_efficient`. When True, removes
the columns of the CIGAR and the difference strings. Both of these are
memory expensive, and might not be used outside of the init of the
class. By default it is True.

Print information about columns at the end. Information about
memory_usage is set to "deep" (does not try to guess an approximate
memory usage value).